+++ packages/react-router/lib/router/utils.ts
@@ -1571,6 +1571,9 @@ export function prependBasename({
   return pathname === "/" ? basename : joinPaths([basename, pathname]);
 }
 
+const ABSOLUTE_URL_REGEX = /^(?:[a-z][a-z0-9+.-]*:|\/\/)/i;
+export const isAbsoluteUrl = (url: string) => ABSOLUTE_URL_REGEX.test(url);
+
 /**
  * Returns a resolved {@link Path} object relative to the given pathname.
  *
@@ -1588,11 +1591,29 @@ export function resolvePath(to: To, fromPathname = "/"): Path {
     hash = "",
   } = typeof to === "string" ? parsePath(to) : to;
 
-  let pathname = toPathname
-    ? toPathname.startsWith("/")
-      ? toPathname
-      : resolvePathname(toPathname, fromPathname)
-    : fromPathname;
+  let pathname: string;
+  if (toPathname) {
+    if (isAbsoluteUrl(toPathname)) {
+      pathname = toPathname;
+    } else {
+      if (toPathname.includes("//")) {
+        let oldPathname = toPathname;
+        toPathname = toPathname.replace(/\/\/+/g, "/");
+        warning(
+          false,
+          `Pathnames cannot have embedded double slashes - normalizing ` +
+            `${oldPathname} -> ${toPathname}`,
+        );
+      }
+      if (toPathname.startsWith("/")) {
+        pathname = resolvePathname(toPathname.substring(1), "/");
+      } else {
+        pathname = resolvePathname(toPathname, fromPathname);
+      }
+    }
+  } else {
+    pathname = fromPathname;
+  }
 
   return {
     pathname,
@@ -1984,7 +2005,8 @@ export class ErrorResponseImpl implements ErrorResponse {
 /**
  * Check if the given error is an {@link ErrorResponse} generated from a 4xx/5xx
  * [`Response`](https://developer.mozilla.org/en-US/docs/Web/API/Response)
- * thrown from an [`action`](../../start/framework/route-module#action)/[`loader`](../../start/framework/route-module#loader)
+ * thrown from an [`action`](../../start/framework/route-module#action) or
+ * [`loader`](../../start/framework/route-module#loader) function.
  *
  * @example
  * import { isRouteErrorResponse } from "react-router";
@@ -2010,7 +2032,6 @@ export class ErrorResponseImpl implements ErrorResponse {
  * @mode data
  * @param error The error to check.
  * @returns `true` if the error is an {@link ErrorResponse}, `false` otherwise.
- *
  */
 export function isRouteErrorResponse(error: any): error is ErrorResponse {
   return (
@@ -2021,6 +2042,12 @@ export function isRouteErrorResponse(error: any): error is ErrorResponse {
     "data" in error
   );
 }
+
+/*
+lol - this comment is needed because the JSDoc parser for `docs.ts` gets confused
+by the star-slash in the `getRoutePattern` regex and messes up the parsed comment
+for `isRouteErrorResponse` above.  This comment seems to reset the parser.
+*/
 
 export function getRoutePattern(paths: (string | undefined)[]) {
   return paths.filter(Boolean).join("/").replace(/\/\/*/g, "/") || "/";