+++ packages/react-router-serve/CHANGELOG.md
@@ -1,5 +1,14 @@
 # `@react-router/serve`
 
+## 7.9.6
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router@7.9.6`
+  - `@react-router/node@7.9.6`
+  - `@react-router/express@7.9.6`
+
 ## 7.9.5
 
 ### Patch Changes